improve calendar component generation
