+++ lib/generate-calendar/lib/templates/page/stylesheet.hbs
