+++ lib/generate-calendar/lib/templates/page/template.hbs
